WIP: CHROMIUM: virtio-wl: Add an ioctl to export dma-buf
A passthrough GPU may have the need to access a virtwl-dmabuf, e.g.
copying the pixels from App's scanout bufffer into a virtwl-dmabuf.

This patch enables the capability to export a virtwl-dmabuf as a
general dma-buf object, so that user space can import the dma-buf to
the passthrough GPU driver and let the buffer be accessed by the GPU.

BUG=None
TEST=Test with a vfio passthrough GPU.

Change-Id: I8a68229793c4d7e5bb5e31fc3d24b5a237723894
Signed-off-by: Tina Zhang <tina.zhang@intel.corp-partner.google.com>