CHROMIUM: arm64: dts: rockchip: Add Kevin rev5
Kevin rev5 is an earlier pre-MP revision. This device tree was not
upstreamed, and likely shouldn't as these units are not publicly
available.

The device tree binding change is included to keep bindings checks happy,
but not in a separate patch for simplicity. Since this won't be
upstreamed, the checkpatch.pl warning is ignored.

BUG=b:197928992
TEST=Build for Kevin; Boot on Kevin DVT2

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: Ia4737b6ca2b30b356fa8a353b2047913813a0696
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3124550
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>