WIP: virtio-gpu: enable PRIME capability to import external buffers
This is a preliminary effort to enable DRI PRIME capability of
virtio-gpu to import buffers from other devices, so that virtual
machine with GPU passed through can offload rendering to the
GPU while display the rendered image via virtio-gpu.

We have a workaround here that we enforce using DMA API in the sg_table
of PRIME buffer to make things work.

BUG=b:267340176
Change-Id: I04ec7e5559d2127363ca41143bf5a88ab151d614